requirements.txt
Django==3.2.25
python-decouple==3.8
django-allauth==0.50.0
django-crispy-forms==1.14.0
crispy-bootstrap4==2022.1
django-countries==7.6.1
Pillow==10.4.0
dj-database-url==0.5.0
psycopg2==2.9.9
stripe==7.12.0